initial commit
This initial commit creates a basic game of hangman (but without any penalties).\nIncluded is a main.go that creates a root cobra command. The cobra command is found in the `cmd` folder. The game state and structures are in the `store` directory. The functions that initalize and play the game are found in the `game` directory